Moved DataAccessException to exception package
