+++ farmstory/src/main/java/farmstory/controller/UserController.java
@@ -5,9 +5,9 @@ import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import farmstory.CountableDAO;
-import farmstory.DataAccessException;
 import farmstory.dao.UserDAO;
 import farmstory.dto.UserDTO;
+import farmstory.exception.DataAccessException;
 import farmstory.service.CountableDefaultService;
 import farmstory.service.CountableService;
 import farmstory.util.ConnectionHelper;
