+++ farmstory/src/main/java/farmstory/exception/DataAccessException.java
@@ -1,4 +1,4 @@
-package farmstory;
+package farmstory.exception;
 
 public class DataAccessException extends Exception {
   public DataAccessException(String message, Throwable e) {
